requirements.txt
# Core ML Frameworks
tensorflow>=2.13.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Data Processing
pywavelets>=1.4.0  # For wavelet transform
scipy>=1.11.0

# MLOps
mlflow>=2.8.0
dvc>=3.0.0
dvc[s3]>=3.0.0  # For S3 support

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data Streaming
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0  # If using MongoDB

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Hyperparameter Tuning
optuna>=3.4.0
ray[tune]>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.0

# Visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Image Processing (for CNN)
opencv-python>=4.8.0
Pillow>=10.1.0

# GUI Framework
streamlit>=1.28.0

# Time Series
statsmodels>=0.14.0

